backend/requirements.txt
uvicorn
fastapi
sqlalchemy
pydantic
certifi
python-jose
passlib
python-multipart
python-dotenv
alembic
requests
psycopg2-binary